Update patch set 3
Patch Set 3:

It lags, for example if you try to play a song in Apollo it will take a lot of time to play/pause, and if you try to change the volume it will also take a lot or the panel will go crazy, until it force closes. In WhatsApp, if you try to play audios that last longer than a minute, it will force close. And i haven't tried any games to report

Patch-set: 3