Update patch set 1
Patch Set 1:

Hmm Michael this flag is actually set to false in your commits, so I wouldn't even need this would I? But devices with multi-color notification LEDs would have to set it to true now or that option will disappear, would it be better to set it to true as default?

Just thinking out loud not sure which is more common, single-color or multi-color.

Patch-set: 1