Update patch set 1
Patch Set 1:

My bad, pressed the wrong button. According to Qualcomm, the 8226 has performance gains with all krait bionic routines except memcpy, which is why they made the kingfisher optimization flag in order to opt in to a9 memcpy as well as the krait CPU variant flag as well as the fact that a7 bionic simply inherits from a15, which is not as optimized as CAF's krait routines. QCOM Bionic optimization only enables the libm optimizations though, so it should probably be tested whether or not both the libm and libc strings from Qualcomm perform better or worse both in bionic_benchmarks as well as the standard fare of applications. I don't have a falcon, but I have friends who do, so I'll look into it.

Patch-set: 1