Update patch set 1
Patch Set 1:

When comparing the diffs of these patches, I noticed that this is not really a revert but a change:
https://github.com/CyanogenMod/android_device_motorola_msm8226-common/commit/d86adc74dba60229d3dba001312e1e15f3a3a697
https://github.com/CyanogenMod/android_device_motorola_msm8226-common/commit/4ddac975ac25948cdbad56b4ba8f5c85fdc614e5

I assume you are aware of it, but it was not obvious for one like me. Would be good to know!

Patch-set: 1
Reviewer: Michael Stucki <10846@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0