Update patch set 1
Patch Set 1:

But why? IMO it's best to keep the structure as close to CAF as possible. This device will still receive updates from Motorola and it should be a lot easier to maintain when that happens.

Maybe just diff this against CAF and move all Motorola changes to one file (e.g. init.target.rc) where possible. Otherwise there would be a ton of other stuff to clean up here as well. And this is coming from someone who's cleaned up this as far as whitespaces and indentations on my devices. What a stupid thing to do - I've already come to regret that. :P

Patch-set: 1