Update patch set 3
Patch Set 3: Code-Review-1

(2 comments)

There was another patch that set the separate voip property to false, which may actually be the correct way to handle this. The property really only needs to be set if there is a real difference in ACDB ID or input path for VOIP.

We probably should check to see if the VOIP paths are actually different (something I probably should have taken some more time to do before commenting on the other patchset), and see if they use a separate ACDB ID as well, otherwise the VOIP property should just be removed.

Patch-set: 3