Update patch set 1
Patch Set 1:

The only response to that I have is this: bugs exist in this device. If someone is willing to help on this, you'd think you'd be grateful for their help - even if they don't own the device.

I am sharing my experiences here to help the community. A response such as this is just downright disgraceful to all of those giving a little bit of their time to help you out with any issue. Elitism should be checked at the door; there is no reason for it here. Besides, if everyone owns the device who "maintains it" then it shouldn't even be a problem for them to pull the required files from the device itself, should it.... meaning they would know they need them.

Patch-set: 1