Update patch set 1
Patch Set 1:

Nothing disrespectful was meant. Anyone that has fixes they think can help should and is welcome to push fixes for review, but the changes should be merged by someone who has actually tested it on the device. That was all that was meant.

The if-exists was removed after issues of nightlies getting produced with no blobs and users updating to a non-booting build.

Patch-set: 1