Update patch set 1
Patch Set 1:

I am aware of the camera issues you mentioned. That is in fact due to the memory issue. I spent a fair amount of time trying to track that one down, but didn't find anything that was an actual fix. As for point 2, I'm not aware of the RIL failure, although I suspect converting a non-GPE to a GPE may actually be the root of that particular problem.

FWIW I am testing on the dual-SIM variant and the single SIM non-GPE unlocked global GSM variant as well.

Patch-set: 1