Update patch set 3
Patch Set 3:

Rabih, it was. I don't know which path is used in stock. Generally it would be a good idea to have two mics recording. Noise cancelling mic is usually on the opposite side of the device than the main mic. You'd get some audio from the direction you're filming as well as the commentary track. :P Could also just try increasing the mic volumes on camcorder-mic-bf path instead of doing this.

Patch-set: 3