Update patch set 1
Patch Set 1:

Accepting that we have to deal with strange firmware combinations isn't something I want to have any support for. It leads to keeping compatibility with every custom anything some guy on the internet has done, which is utterly impractical. This isn't a supported use case in the same way that custom kernels/ramdisks isn't a supported use case.

Patch-set: 1