Update patch set 3
Patch Set 3:

Can you check the log while AudioFx is active? If it's working you should see quite a lot of log statements generated. If the effects are working but not to your liking that's a different concern. I'm mostly worried that there is something actually broken in the code.

Patch-set: 3