Update patch set 1
Patch Set 1:

Since the issue happens pretty randomly and you can´t replicate it at any time (you just have to wait), it´s hard to confirm it for sure.

On the other hand there are currently hundreds of people using CM 12 with disabled KSM either through a modified kernel or through using a custom build of CM.
If it didn´t work, there would be someone who would have told us by now, like with all the previos attempts to solve this.

Thanks for submitting this and merging it.

Patch-set: 1