Update patch set 3
Patch Set 3:

titan has working device tree now...so we have to figure out  the changes to be made to msm8226-common repo for titan without breaking other two variants.
so device specific stuff to be moved out to specific trees ..so that we can keep this tree clean.

Patch-set: 3