Update patch set 5
Patch Set 5: Code-Review+1

I measured both cpuinfo and meminfo with and without the patch and this is the outcome:
http://visi0nary.de/wp-content/misc/dumpsys_meminfo_w_patch.txt
http://visi0nary.de/wp-content/misc/dumpsys_meminfo_without_patch.txt
http://visi0nary.de/wp-content/misc/dumpsys_cpu_info_w_patch.txt
http://visi0nary.de/wp-content/misc/dumpsys_cpuinfo_without_patch.txt

All measurements were captured after a fresh boot and after I started Firefox, Spotify, TextSecure and the Settings App.
Obviously the OOM is adjusted to 86000 whereas in without the patch it's 325000 (last line). Furthermore, iowait in cpuinfo is nearly at 10% due to frequent reloading and resulting IO. Apart from that we can clearly see that "cached PSS" is nearly 6x higher than without the patch but "cached" is decreased in the "Free RAM" line. Last but not least RAM status is "moderate" vs. "critical".

I'm currently looking into the "<integer name="config_extraFreeKbytesAbsolute">0</integer>" or rather if this is really needed - it doesn't seem to make any difference though.

Patch-set: 5