Update patch set 2
Patch Set 2:

I was working at lp kernel but I'm unable to build media stuff and I have no clue for fix it .
I don't know is someone else is working at it.
We ( titan ) are waiting for it for start nightlies but seems we will "never" start them xd.

Patch-set: 2