Update patch set 10
Patch Set 10: -Code-Review

LOL, I never left display settings when I tested it. :P No lags with the older version of kcal I was previously using (linked in the second comment here) so it must be an issue with the new kernel implementation.

Also IMO it would make more sense to do this after we've moved to unified 'lollipop—5.0.2-release' based kernel.

Patch-set: 10