Update patch set 2
Patch Set 2:

That´s the error. Camera always used to work for a while with KSM on (sometimes up to half a day) and then stopped working.

If anybody was interested, this line would always appear in dmesg after taking a shot with broken camera on the old CM12 kernel:
<3>[ 5373.045069,1] platform 16.qcom,ion-heap: Fail to allocate buffer

Patch-set: 2