Update patch set 6
Abandoned

Sorry for the trouble, its a late night for me. I'll figure it out when I get some free time. 

I'm basically having an issue with my display going below 10 when it is clearly defined to be 10 and assumed this would be the way to go. My original intention by setting it as 3 would result in the dimness never going below 3.

Patch-set: 6
Status: abandoned