Update patch set 7
Patch Set 7:

Actually I am not sure if it affects only Whatsapp or also other apps. I remember that this was mentioned somewhere in a forum.

However, I tried right now to make a VOIP call using the builtin dialer as well as with Zoiper, and both works even without this change...

Patch-set: 7