Update patch set 2
Patch Set 2:

And BTW while on the subject, there's a Kconfig option in prima makefile that's not right when built as part of the kernel (it's fine when prima is built as part of CAF build like Motorola does). Also defconfigs need the corresponding change and that option should also be enabled (per https://github.com/MotorolaMobilityLLC/vendor-qcom-opensource-wlan-prima/blob/MMI-LPB23.13-35/Kbuild#L44). I don't have the CM kernel synced so I can't be bothered to submit those changes.
https://www.codeaurora.org/cgit/quic/la/platform/vendor/qcom-opensource/wlan/prima/commit/Kconfig?h=LA.BF.1.1.2_rb1.24&id=377314d17263630b114523845e3d9ffae29d960f

Patch-set: 2