Update patch set 2
Patch Set 2:

Errm, that's exactly what's wrong with it. Even if you tried to enable CONFIG_CONFIG_ENABLE_LINUX_REG it wouldn't do anything because the code itself is ifdef'ed with CONFIG_ENABLE_LINUX_REG. And that option _should_ be enabled because it's enabled by default in CAF/Moto builds.

For reference, bacon:
https://github.com/CyanogenMod/android_kernel_oneplus_msm8974/commit/f85e35bb8a71a9e7b4d9951784988177381a7863#diff-024229054b227eb6eca7f2f49dac938aL48
https://github.com/CyanogenMod/android_kernel_oneplus_msm8974/blob/cm-12.1/arch/arm/configs/cyanogenmod_bacon_defconfig#L3093

Patch-set: 2