Update patch set 1
Patch Set 1:

Luca and Lukasz, could you try this patch with the addition of

 <integer name="config_proximityCheckTimeout">100</integer>

and see if the feature works properly on your Moto G variants?
I'm currently using it and it works great, so I can't see why not merge it, but it would be nice to know if this can go in msm8226-common or has to be moved in falcon.

Patch-set: 1