Update patch set 3
Patch Set 3:

We need to verify the WHITE_LED_FILE vs RGB_CONTROL_FILE thing.

I guess that the different entries in the dst actually correspond to the same LED (probably Motorola planned to have different leds, as shown by some pre-production phones, but then changed their mind).

If you can, try to see if the result of the following commands is what I expect it to be:

 echo ff0000 100 100 1 1 > /sys/class/leds/rgb/control # blink very fast
 echo 00000 100 100 1 1 > /sys/class/leds/rgb/control # led off
 echo ff0000 100 100 1 1 > /sys/class/leds/rgb/control # blink very fast
 echo 255 > /sys/class/leds/white/brightness # static light
 echo 0 > /sys/class/leds/rgb/control # led off
 echo 255 > /sys/class/leds/white/brightness # static light
 echo ff0000 100 100 1 1 > /sys/class/leds/rgb/control # blink very fast
 echo 0 > /sys/class/leds/rgb/control # led off

Patch-set: 3