Update patch set 1
Patch Set 1:

According to the defconfig, thea uses CT406, same as falcon. The sensor is disabled on actual suspend and not when the screen is turned off (using the MMI panel notification). I see no lines there suggesting that the sensor is either turned on or off, so unless you changed the debug mask, it probably means you didn't enable the feature from the settings.

Patch-set: 1