Update patch set 1
Patch Set 1:

Sorry, I saw CT406 enabled and thought it used the same sensor.

Just a note. The input subsystem filters out all the consequent events with the same value. This means that if the kernel sent one value to userspace, the next time it won't report it unless it changed, so if it has to send "sensor covered" two times, it might wont. I'd make sure this doesn't happen with your sensor.

Patch-set: 1