Update patch set 1
Patch Set 1:

The interval is either too short or it's happening what I said in my previous comment. Anyway, this means this should be moved to falcon.

> Also sometimes I get this weird PM spam http://pastebin.com/hvHit1wp

The IRQ of the sensor is configured as wakeup source on suspend (enable_irq_wake()) if the proximity sensor is enabled. If you have something that uses the proximity sensor to do some fancy stuff, the device wakes up whenever something happens near the sensor (I guess this is the case if you have some gesture enabled).

Patch-set: 1