Update patch set 3
Patch Set 3:

Not completely related, but I was informed that there are some issues with OMX. In particular, libOmxVenc.so seems to be the faulty component and it prevents apps from encoding video. Is there any reason not to momentarily switch to prebuilt libraries (libOmxVenc.so at least, I don't know if there are other problems)?

Patch-set: 3