Update patch set 3
Patch Set 3:

Those are debug lines that I enabled.

Anyway, we simply have to get rid of that extra size, not make the buffer bigger. If you look at Motorola's sources, the extra size is not included in VENUS_BUFFER_SIZE(). This header is used to build libOmxVenc.so, so when we build it using our kernel sources, we have those extra 8192 bytes. This means that the prebuilt lib is not allocating bigger buffers, it's simply doing the right math.

(Yes, we can move the conversation elsewhere if you prefer. You can also contact me by email.)

Patch-set: 3