Update patch set 2
Patch Set 2: Code-Review+1

I think this is a nice feature and on JellyBean or KitKat the led behavior was exactly this.
It works fine unless the phone is rebooted in Low Power Mode when it's already fully charged. In this case the led takes a while to power on

Patch-set: 2