Update patch set 4
Patch Set 4:

Last time I checked, the led didn't power on without setting 255; but I'll check another time it. Turning off the led after a while would make the feature useless. The led is useful to know when the phone is charged without turning on the screen; it is a faster check in other words. "battery-full" will turn off the led when the charger is disconnected; in this way we are sure that users know when the device is charged.
If you are wondering about consumptions; the phone is in low power mode, and the led is on only when the charger is inserted, even if the battery level drains a bit, it's instantly recharged since the charger is still connected.

Patch-set: 4