Update patch set 4
Abandoned

As I already said: "I think this is a nice feature and on JellyBean or KitKat the led behavior was exactly this."
Our prebuilt binary doesn't do anything similar and since it is prebuilt we can't edit it.
Anyway the led turns on even without setting the brightness, it just takes a bit more. I don't have time enough to investigate at the moment, so I'll abandon this.

Patch-set: 4
Status: abandoned