Update patch set 4
Patch Set 4:

> Anyway the led turns on even without setting the brightness, it just takes a bit more. I don't have time enough to investigate at the moment, so I'll abandon this.

FYI, the trigger turns the LED on when the battery is full, not when it's at 100%, which are two different things.

I have a question. According to the history, the source built charge_only_mode was removed because USB charging doesn't work with it. I can't replicate the issue. I remember having issues with stock ROMs until a certain update (the phone simply rebooted), so maybe the updated kernel sources/bootloader fixed the issue. I also modified the code to improve the compatibility with my falcon. Any other reason (aside from better graphics, which can be changed) to keep using the prebuilt one?

Patch-set: 4