Update patch set 4
Patch Set 4:

I tried different chargers, even Apple ones, but I've managed to have issues with a computer, which simply makes my phone reboot continuously.

A quick string search of the prebuilt binary shows that both 'online' and 'presence' must be used to know whether the usb cable is insterted or not and this indeed solved the reboot problem.

As for the question mark, that might be a matter of adding a delay.

Well, this is not important now, I was just wondering why we were still using the prebuilt binary.

Patch-set: 4