Update patch set 2
Patch Set 2:

How is this even working properly? I was waiting for this to be merged, but I had the patch ready, which actually depends on http://review.cyanogenmod.org/#/c/106399/ (where I also mention this).

The brightness is encoded in the alpha channel of the ARGB color and must be extracted. Our current liblight is simply discarding the alpha channel and using the color of the LED to change the brightness.

Patch-set: 2