Update patch set 2
Patch Set 2:

Sorry, I was truly wondering how this was working and found the answer.

We are currently not masking state->color, and we use "%6x" to print it, expecting to print RGB. However, now state->color is not RGB, but ARGB and doing "%6x" will print ARG instead of the expected RGB. Since the alpha channel now replaces the red channel, it effectively controls the brightness.

Patch-set: 2