Update patch set 4
Patch Set 4:

> The brightness values is always 0

There's a bug somewhere in the preferences, it might be related to the fact that we have config_multiColorNotificationLed.

What happens is that while moving the slider, A=0xFF, R, G and B instead are all equal to the brightness level.
Once you close the slider to change the brightness, however, the correct brightness value is saved and the next time the LED is turned on, the saved brightness level will be used. (Try to lower the brightness and then try to edit the LED preferences)

Patch-set: 4