Update patch set 1
Patch Set 1:

This is mostly relevant for offline charging while the phone is plugged to a Windows machine.

I noticed this while testing the source built charge_only_mode, which kept showed a question mark. The prebuilt charge_only_mode didn't show any error, but the device wasn't charging.

PS: I looked a bit more into this all charge_only_mode and I can say (not 100% sure though) that it can't affect offline charging. The only problems with are the paths to be check to determine whether the device should be turn off or not (I think I fixed this).

Patch-set: 1