Update patch set 2
Patch Set 2:

It's still not clear to me how to reproduce the problem.

Do I have to accept the call?
What's the exact sequence of actions, is it known?
Does it happen randomly and appears to happen only after receiving a call?
Are there other situations in which this might happen?

AFAIK, the proximity sensor kicks in only after a call had been accepted.

PS: it's not a matter of dark vs bright, but close vs far.

Patch-set: 2