Update patch set 1
Patch Set 1:

Oh, and yes, 'charging' is not used for offline charging, but since the purpose of this config is charging the phone, why not use the same config used for offline charging? (Plus, the 'charging' 'functions' doesn't exist)

When I set the old charging mode, Windows keep showing "Unrecognized device connected" (or whatever the message is). This config fixes the problem and the phone starts charging with no issues.

Patch-set: 1