Update patch set 1
Patch Set 1:

Oh, I thought I was explaining myself poorly. 

The reason I thought so is because the function and id you changed "charging" to is likely the same as the "mass_storage" one. Additionally, when you are in off mode charging, the upper layers of Android that normally connect mass storage to the USB gadget are not active, so it does nothing except provide the id.

Qualcomm is setting "on charger: setprop sys.usb.config mass_storage" on their default init scripts as well, so this is not that weird of a thing to do.

To answer your other question, "charging" should be used for the charging only, no MTP, no mass storage case in Android.

Patch-set: 1