Update patch set 14
Patch Set 14:

I wasn't saying it didn't 'work'. It's picked up where it was last defined. LOCAL_PATH variable is reused and redefined hundreds of times when makefiles are parsed during the build. If it was last defined in say motorola/msm8226-common/liblight/Android.mk, you'd be screwed and your build would explode. If it was last defined in BoardConfig.mk and now you changed it to VENDOR_PATH, that would also kind of make a difference because now it would be where you last defined LOCAL_PATH before that. It's likely from motorola/msm8226-common/Android.mk, though, which is why it works. Every makefiles that uses LOCAL_PATH usually sets 'LOCAL_PATH := $(call my-dir)' at the very beginning of that makefile.

Patch-set: 14