Update patch set 1
Patch Set 1:

I recently found the time to look into PowerHAL and found some issues. The most important was the lack of "profile changed" hints. This becomes obvious if you set the performance profile since it floods the logcat with error messages.

I tried to use liqc-opt.so, but I think its capabilities are pretty limited and decided to write a custom HAL. Not even Motorola is using it. Plus, the QCOM HAL was simply boosting the CPU for 3 seconds on every user interaction, so nothing we can't do without libqc-opt.so and I don't really like to use precompiled binaries for no reason.

I dropped cpu_boost because interactive doesn't really need it and actually boosting the CPU through boostpulse changes the codepath a bit, so it's not entirely the same.

I've been using something similar to this for a while, but I did some changes recently, so it would be nice if you could test this.

I'm not sure yet about a couple of settings, but this configuration had worked pretty well so far.

Patch-set: 1