Update patch set 2
Patch Set 2:

As soon as audio bars are shown, be it in the notification panel (AudioFX toggle), in the lockscreen or whatever, the usecase changes from compress-offload-playback to deep-buffer-playback. This switch also causes an annoying click noise and can't get back to compress-offload-playback.

hardware/qcom/audio-caf/msm8974/
http://pastebin.com/raw.php?i=wwqxVySK

If you get lines such as:

> V/audio_hw_primary(12307): out_write: writing buffer (3840 bytes) to pcm device

then the CPU is processing the audio.

Patch-set: 2