Update patch set 2
Patch Set 2:

> I've never seen errors in the kernel log, probably because the usecase always changes for me

Regarding this, I had to enable audio effects making sure that no audio bars were shown and that offloading was possible. I guess that the difference between visualizer and the other effects is that userspace needs something in return, so in that case it's clear when something goes wrong on the kernel side (I haven't looked at the sources, but it makes sense).

Anyway, the main problem with the visualizer effect is that it's nearly everywhere. It would be nice to have a setting which automatically disables audio bars for compress-offload.

IMO we should hide the AudioFX toggle by default (easy), disable audio bars in the lockscreen (no overlay for this I think) and maybe disable audio bars in Eleven as well (same). Because otherwise offloading becomes nearly impossible considering the 2 hours interval. This would actually be fine if it's wasn't for the annoying click sound that happens when offloading is aborted mid-track.

Patch-set: 2