Update patch set 2
Patch Set 2:

@Gabriel: On titan offloading works without problems as long as no effects are used. That's why in this commit you see all offload_bundle (hw effects) disabled. This allows for userspace to properly switch between 'no effects' + compress-offload and audio effects + deep-buffer, based on the AudioFX toggle.

Just to make sure I explained myself correctly. No effect from the offload_bundle seems to work it's not just about the visualizer. What is happening and what I think you noticed is, that visualizer is the only effect that returns to userspace therefore if it fails userspace aborts (invalidates) offload. Userspace then proceeds to use deep-buffer + sw effects and the users thinks that everything is working normally but behind the scenes it is just because we are using deep-buffer now instead of offload.

Patch-set: 2