Update patch set 2
Patch Set 2:

Yes, you explained the problem and I understood what you are trying to do. And for what is worth, the problem is the same on falcon.

I was just saying that it's pretty hard to avoid the visualizer effect on a default installation. If you are a new user, what you notice is that from time to time music playback momentarily stops (offload aborted) and then re-starts (switch to deep-buffer). You probably don't even know what offload is and you might find the click noise annoying. So I was suggesting to minimize the chances of aborting audio offload by hiding all the audio bars that are enabled by default.

This problem is even mentioned here: http://review.cyanogenmod.org/#/c/70805/

Patch-set: 2