Update patch set 4
Patch Set 4:

Generally, when you are in off mode charge, most devices use the same VID/PID as mass_storage, that's all I've been trying to say. The easiest way to do that is to just set the USB composition mode to mass_storage.

The "charging" composition mode is for the case where the device is plugged into a computer, but MTP/PTP/UMS is not active.

This current patchset appears to be fine with regard to both of these.

If I had to guess, it may be because the hanging computers have USB drivers of some sort installed? I'm unsure if using the CAF VID/PID is better here.

Patch-set: 4