Update patch set 1
Patch Set 1:

> Titan and thea have the exact same CPU as falcon - still a quad Cortex-A7 at 1200 MHz.

Yes, I know. However, Motorola is using an higher hispeed_freq for falcon (this could be just a mistake). I also did a side-by-side comparison between my falcon, a titan and a peregrine and the average CPU freq of my falcon is higher (-> higher average cpu load) and it appears to be less responsive than the other two.

It's worth mentioning that those two other devices are running a stock firmware, but I had the same impression even when my device was running a stock firmware.

I might be wrong, these are just impressions and there are many factors to take into account, but it's also true that not all the chips perform in the same way and it's not all about frequency.

Patch-set: 1