Update patch set 5
Patch Set 5: Code-Review+2 Verified+1

I think there isn't a clean and easy way to get the properties defined for offline charging. This means computers will see a generic "Android" rather than "Moto G XXX", but this is not going to change the way the phone detected. Moreover, the phone will not show up in file managers (which is where having a proper name is useful) for this particular case, so hacking init just for this is not even worth it.

So let's get this charge only mode fixed.

Patch-set: 5
Reviewer: Gabriele M <13897@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>