Update patch set 1
Patch Set 1: Code-Review+1

Afaik there was some issue with "missing phone number" back early cm-11.0 but I'm not sure if it was about "Settings->About->Status" which my carrier doesn't support or incoming calls. Not sure if this issue is a case now but source built rild/libril.so seems to work correctly.

Patch-set: 1
Label: Code-Review=+1