Update patch set 1
Patch Set 1:

hw visualizer was disabled some time ago. I did read the comments regarding the issues and therefore never bothered testing what would happen if I enable it. However it seems to be working as intended.

I want to ask if the problems from the past were due to visualizer_hw not 'aborting' offload playback (intended) which in turn lead to userspace NOT opting for deep-buffer playback. As we know deep-buffer/low-latency playback chooses sw effects, which gives us audio playback with effects.

As we also know offlaod_playback currently does not work with effects and therefore we disabled them. However if the user choose to enable the visualizer (which has nothing to do with audio effects) userspace will abort the offload (even for non-effect playback). This is because it can not 'listen' to the DSP output and display it as the visualizer bars onscreen.

Patch-set: 1