Update patch set 1
Patch Set 1:

Something is not clear to me.

Is the lack of the HW visualizer causing any issue? I mean, is there any situation in which the lack of the HW visualizer is preventing audio offload?

We know that the HW visualizer doesn't work. As soon as userspace tries to use it, it immediately fall backs to the SW visualizer. And as far as I know, this switch happens only when userspace tries to actually show audio bars.

This means the HW visualizer is never really used and having it makes no difference. Am I wrong?

Patch-set: 1