Update patch set 1
Patch Set 1: Code-Review+1

Kernel support is really required for the always on mode, which we won't enable.

Also, the doze states require command mode panels, while we have video mode panels (falcon at least). I'm not sure it's possible to get it working with video mode panels, AFAIK there's no code available, but the thing is that a pulsating display shouldn't cause a major power consumption.

+1 from me, but I'm open to discussion.

Do you know if Ambient Display is enabled by default? Because if it is, I'd disable it by default as Motorola is doing it.

Patch-set: 1