Update patch set 1
Patch Set 1:

In our chmod, '-L' is implicit and actually specifying it makes chmod exit with an error. So part of the original code is broken.

I guess that the reason why exec was here used was that the in-built commands behave differently.

This makes me wonder one thing: is ueventd traversing the links?

Patch-set: 1