Update patch set 2
Patch Set 2:

Yeah, I really don't know.

On one side it's true that we just have a copy, but it's also true that this will bring some extra stuff. For instance, I'm not sure what PersistentStorage is used for, I didn't look into this at all, but I know we have /persist, so I expect that cmhw will try to use it.

Patch-set: 2