Update patch set 2
Patch Set 2:

I don't remember why I kept the absolute path, it was like this in the original script. If I'm not wrong, /system/bin/dd and /system/xbin/dd used to be different and the latter was a bit more powerful, so probably that's why I kept the full path. However, it shouldn't matter in this case since we are doing a simple dump, so I'd simply use dd.

Patch-set: 2