Update patch set 3
Patch Set 3:

Using the touchscreen as example is not OK because Motorola implemented a custom mechanism to suspend it as soon as the screen is turned off.

This flag allows the device to suspend when the screen is turned off because of the proximity sensor (as the description says). If I'm not mistaken, it changes the value of /sys/power/autosleep and maybe some other things.

The reason this flag exists is that the proximity sensor must be able to wake up the device, which is not necessarily true. I know the proximity sensor of falcon can.

Patch-set: 3
Reviewer: Gabriele M <13897@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0