Update patch set 6
Patch Set 6:

Thanks Gabriele. 

Just leaving this here: on titan what I could measure, when covering prox sensor in call, was that it takes about 150ms for the sensor to calibrate and report a value. When waking (ambient display) only calibration sequence is reported but it does not get to the point of reporting a value. I tried to tweak the 'pulse_duration_in' values, but this without any luck. 'TIMEOUT_DELAY_MS=500' in DozeService.java is well set to 500ms which is enough time ...

Patch-set: 6