Update patch set 1
Patch Set 1:

This looks OK to me.

The HAL doesn't report which sensor is used, unlike the stock one, so switching to bmc055 should make any different from that point of view.

If according to the HAL bmc150 is bma255+bmm150 and we know that titan has bma255+bmm050, I'd switch to bmc055, even if the sensor is actually a bmc150.

I'd maybe add a comment in the code to clear things up.

Patch-set: 1
Label: Code-Review=0