Update patch set 1
Patch Set 1: Code-Review-1

Bosch provides a combination of the BMA2x2 + BMM050 APIs for the BMC150 and BMC056 and none for the BMC050. So actually judging by their commit msg (see Luk's post), I do now believe they put in a BMC150 but names got mixed because of the definitions in the API.

For the record everything was kinda missleading. This is what the kernel log reports and therefore confused me:

[    0.894032,0] [N]<5><BMP280><bmm_probe><1320>Bosch Sensortec Device bmm050 detected: 0x13
[    0.894324,0] input: bmm050 as /devices/virtual/input/input3
[    0.904731,0] [N]<5><BMP280><bmm_probe><1442>sensor bmm050 probed successfully
[    0.904970,0] i2c-core: driver [bmm050] using legacy suspend method
[    0.905048,0] i2c-core: driver [bmm050] using legacy resume method

Let's leave it as it is, I see no gain in using the 'magic numbers' for bmm050...

Patch-set: 1
Label: Code-Review=-1