Update patch set 2
Patch Set 2:

Is there any practical difference, aside from the one I mentioned in the commit?

Appending the tree blob seems to work just fine and in both cases, it's the bootloader that gets it, because unlike for BUILD_ARM_APPENDED_DTB_IMAGE, the kernel does nothing. So I guess in the end it doesn't matter much where the dtb is, as long as the bootloader finds it.

I find annoying having to flash the kernel to test changes to the device tree, rather than just doing a simple 'fastboot boot'.
It's not something I do every day, but it's still a nice option to have.

So, is there any actual difference that I'm missing? I haven't really looked into dtbtool, so I really don't know, I just tested this and it worked just fine.

Patch-set: 2