Update patch set 9
Patch Set 9:

I looked better into this. The previous scheduler was IMO wrong.

The animation takes a lot. If you set 1s, 5s and 10s, you end up with the screen turning on and off with no pause, because the time for the three pulses is greater than 15s, at least with the default configuration. With the tweaked one if it wasn't 15s, it was really close to it (too lazy to do the math). The default one is fine IMO.

Patch-set: 9