Update patch set 3
Patch Set 3:

> > AFAIK minDelay/maxDelay don't really do anything. Te HAL should
 > > ensure they are respected, so in the end, this should have no
 > > noticeable effect for the user.
 > >
 > > Anyway, isn't the commit message saying the opposite of what's
 > > actually done?
 > 
 > No. The original value was 20000 (1 event every 20000 microseconds,
 > or 50hz), the new value is 50000 (1 event every 50000 usec, 20hz).

Also, getMinDelay/getMaxDelay _are_ part of the public sensors API (and that's the API CTS uses for tests). An app _can_ legitimately look at those values and assume "OK, if this can give me 50hz, I want 25 events per second", use that value in the SensorManager.registerListener() call (third argument to that is the delay) and then it'd be strangled at 20.

Patch-set: 3