Update patch set 3
Patch Set 3:

> No. The original value was 20000 (1 event every 20000 microseconds, or 50hz), the new value is 50000 (1 event every 50000 usec, 20hz).

Right, I saw 50hz/20hz, but read 50ms/20ms, probably because I looked at the change first and then read the commit message.

 > Also, getMinDelay/getMaxDelay _are_ part of the public sensors API (and that's the API CTS uses for tests). An app _can_ legitimately look at those values and assume "OK, if this can give me 50hz, I want 25 events per second", use that value in the SensorManager.registerListener() call (third argument to that is the delay) and then it'd be strangled at 20.

Yeah, what I meant to say was that changing this maxDelay and minDelay doesn't actually clamp the rate. Saying that apps are not affected by it is definitely wrong though.

Thanks for the explanation, I was in a rush and couldn't see everything properly (and that's why I didn't merge this).

Patch-set: 3