Update patch set 2
Patch Set 2: Code-Review+1

> I'm not sure of this yet. I didn't have problems with the previous
 > audio_policy.conf and only recently moved to this one.

I'm still trying to understand how changing the order of the flags for the bitfield could have any implication. But regardless I trust CAF on this one...

Anyway, I have rebooted my phone about 30 time the past 2 days and not come across the issue. I guess we could merge this before the first people encounter the 'bug'.

For the long term, the issue 'may' be in the hal but gone unnoticed because most the devices use a config very close to CAF... and I think we should too. At least for the stuff we support at kernel level. This goes both for audio_policy and mixer_paths. I.e I've seen 'afe-proxy-record' usercase entries in the logcat (just be4 callrecord started), but it's not defined in  mixer_paths.xml

Patch-set: 2