Update patch set 2
Patch Set 2:

Is this making any difference? Because I barely notice it.

The difference is less than what I thought, something like 1 or 2 minutes compared to no pre-dexopt at all, so I'm still not sure about this change. The thing is that this change makes releasing single framework jars more complicated and I did that in the past to test changes on variants I don't own.

Full pre-dexopt really cuts down the time to boot and I'd be ok with that, but I need to look better at the gapps thing. I really don't want to break current installations.

Patch-set: 2