Update patch set 2
Patch Set 2:

The main difference is seen on encrypted devices, without something like this it takes a long time to show the pin pad every single time you boot. I haven't looked at the log tbh but it looks to me that it dexopts all the core stuff on every single boot because cache is on data, and data is encrypted.

Patch-set: 2