Update patch set 2
Patch Set 2: Code-Review+1

Well, I'm okay with this.

I really would like to enable the full optimization, but we'd need to implement some hack for those who flashed gapps packages 200MB+ big. I did some tests and it seems that the restore fails silently and the the file in addons.d is not created, which is definitely not good.

Patch-set: 2