Update patch set 1
Patch Set 1:

I was thinking about this aswell, but to me it is kind of an 'unneeded' feature as accidentally pressing the power button on titan/thea when it is in pocket (unless you wanted to press vol keys to skip music) is very difficult. I guess it is more handy for devices with physical button(s) on front (Samsung).

Patch-set: 1